# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Estecapelli API — Preevaluación capilar
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    post:
      operationId: createPatient
      tags: [patients]
      summary: Create a new patient pre-evaluation (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResult"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{token}:
    get:
      operationId: getPatient
      tags: [patients]
      summary: Get patient pre-evaluation by token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePatient
      tags: [patients]
      summary: Update patient pre-evaluation by token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResult"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{token}/photos:
    get:
      operationId: getPatientPhotoStatus
      tags: [patients]
      summary: List private clinical photo metadata for the token holder
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Photo state without image bytes or object URLs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPhotoStatusResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: uploadPatientPhoto
      tags: [patients]
      summary: Upload a private original clinical photo for the token holder
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: x-photo-key
          in: header
          required: true
          schema:
            type: string
        - name: x-photo-source
          in: header
          required: true
          schema:
            type: string
            enum: [camera, upload]
        - name: x-photo-width
          in: header
          schema:
            type: string
        - name: x-photo-height
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Draft original stored privately
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPhotoStatus"
        "400":
          description: Invalid photo input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: discardPatientPhotos
      tags: [patients]
      summary: Discard all draft photos for a patient pre-evaluation (public, token-authorized)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Draft photos discarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{token}/photos/{photoId}/confirm:
    post:
      operationId: confirmPatientPhoto
      tags: [patients]
      summary: Confirm a draft clinical photo for the token holder
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Confirmed photo metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPhotoStatus"
        "404":
          description: Draft photo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{token}/photos/{photoId}/adjusted:
    post:
      operationId: createPatientAdjustedPhoto
      tags: [patients]
      summary: Store an adjusted derivative without replacing the original
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: photoId
          in: path
          required: true
          schema:
            type: string
        - name: x-edit-params
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
          image/png:
            schema:
              type: string
              format: binary
          image/webp:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Adjusted derivative metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPhotoStatus"
        "404":
          description: Original photo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /patients/{token}/photos/{photoId}:
    delete:
      operationId: discardPatientPhoto
      tags: [patients]
      summary: Discard one draft or superseded clinical photo for the token holder
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: photoId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Draft photo discarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleOk"
        "404":
          description: Draft photo not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    PatientPhotoStatus:
      type: object
      required: [id, key, label, status, source, mimeType, sizeBytes, sha256, createdAt, hasOriginal, hasAdjusted]
      properties:
        id:
          type: string
        key:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [draft, confirmed, superseded, discarded]
        source:
          type: string
          enum: [camera, upload]
        mimeType:
          type: string
        sizeBytes:
          type: number
        sha256:
          type: string
        createdAt:
          type: string
          format: date-time
        confirmedAt:
          type: ["string", "null"]
          format: date-time
        hasOriginal:
          type: boolean
        hasAdjusted:
          type: boolean
        editParams: {}
    PatientPhotoStatusResult:
      type: object
      required: [ok, photos]
      properties:
        ok:
          type: boolean
        photos:
          type: array
          items:
            $ref: "#/components/schemas/PatientPhotoStatus"
    LeadStatus:
      type: string
      enum: [nuevo, incompleto, listo, contactar, agendado, cerrado]
    LeadSummary:
      type: object
      required: [id, token, createdAt, updatedAt, name, phone, status, photoCount]
      properties:
        id:
          type: string
        token:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        name:
          type: string
        phone:
          type: string
        documentId:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        age:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/LeadStatus"
        consent:
          type: boolean
        photoCount:
          type: number
        photoKeys:
          type: array
          items:
            type: string
        hairLossTime:
          type: ["string", "null"]
        pattern:
          type: ["string", "null"]
        previousTreatment:
          type: ["string", "null"]
        norwood:
          type: ["string", "null"]
        appointmentAt:
          type: ["string", "null"]
        isDemo:
          type: ["boolean", "null"]
...
    PatientInput:
      type: object
      required: [name, phone, consent]
      properties:
        name:
          type: string
        phone:
          type: string
        documentId:
          type: string
        email:
          type: string
        age:
          type: string
        city:
          type: string
        hairLossTime:
          type: string
        pattern:
          type: string
        previousTreatment:
          type: string
        symptoms:
          type: string
        surgeryHistory:
          type: string
        consent:
          type: boolean
        marketingConsent:
          type: boolean
    PatientResult:
      type: object
      required: [ok, lead]
      properties:
        ok:
          type: boolean
        lead:
          $ref: "#/components/schemas/LeadSummary"