# lib/api-zod/src/generated/types/patientInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Estecapelli API — Preevaluación capilar
 * OpenAPI spec version: 0.1.0
 */

export interface PatientInput {
  name: string;
  phone: string;
  documentId?: string;
  email?: string;
  age?: string;
  city?: string;
  hairLossTime?: string;
  pattern?: string;
  previousTreatment?: string;
  symptoms?: string;
  surgeryHistory?: string;
  consent: boolean;
  marketingConsent?: boolean;
}