# lib/api-zod/src/generated/types/patientPhotoStatusResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Estecapelli API — Preevaluación capilar
 * OpenAPI spec version: 0.1.0
 */
import type { PatientPhotoStatus } from './patientPhotoStatus';

export interface PatientPhotoStatusResult {
  ok: boolean;
  photos: PatientPhotoStatus[];
}